// sdk/src/contracts/CollectionContract/types.ts
export type FeatureInterfacesMap = {
  ICedarFeaturesV0?: FeatureInterface<ICedarFeaturesV0>;
  ICedarVersionedV0?: FeatureInterface<ICedarVersionedV0>;
  ICedarVersionedV1?: FeatureInterface<ICedarVersionedV1>;
  ICedarVersionedV2?: FeatureInterface<ICedarVersionedV2>;
  IMulticallableV0?: FeatureInterface<IMulticallableV0>;
  ICedarAgreementV0?: FeatureInterface<ICedarAgreementV0>;
  ICedarAgreementV1?: FeatureInterface<ICedarAgreementV1>;
  IPublicAgreementV0?: FeatureInterface<IPublicAgreementV0>;
  IPublicAgreementV1?: FeatureInterface<IPublicAgreementV1>;
  ICedarUpdateBaseURIV0?: FeatureInterface<ICedarUpdateBaseURIV0>;
  ICedarUpgradeBaseURIV0?: FeatureInterface<ICedarUpgradeBaseURIV0>;
  IPublicUpdateBaseURIV0?: FeatureInterface<IPublicUpdateBaseURIV0>;
  ICedarNFTIssuanceV0?: FeatureInterface<ICedarNFTIssuanceV0>;
  ICedarNFTIssuanceV1?: FeatureInterface<ICedarNFTIssuanceV1>;
  ICedarNFTIssuanceV2?: FeatureInterface<ICedarNFTIssuanceV2>;
  ICedarNFTIssuanceV3?: FeatureInterface<ICedarNFTIssuanceV3>;
  ICedarNFTIssuanceV4?: FeatureInterface<ICedarNFTIssuanceV4>;
  IPublicNFTIssuanceV0?: FeatureInterface<IPublicNFTIssuanceV0>;
  IPublicNFTIssuanceV1?: FeatureInterface<IPublicNFTIssuanceV1>;
  ICedarSFTIssuanceV0?: FeatureInterface<ICedarSFTIssuanceV0>;
  ICedarSFTIssuanceV1?: FeatureInterface<ICedarSFTIssuanceV1>;
  ICedarSFTIssuanceV2?: FeatureInterface<ICedarSFTIssuanceV2>;
  ICedarSFTIssuanceV3?: FeatureInterface<ICedarSFTIssuanceV3>;
  IPublicSFTIssuanceV0?: FeatureInterface<IPublicSFTIssuanceV0>;
  IPublicSFTIssuanceV1?: FeatureInterface<IPublicSFTIssuanceV1>;
  ICedarPausableV0?: FeatureInterface<ICedarPausableV0>;
  ICedarPremintV0?: FeatureInterface<ICedarPremintV0>;
  ICedarLazyMintV0?: FeatureInterface<ICedarLazyMintV0>;
  ICedarNFTMetadataV1?: FeatureInterface<ICedarNFTMetadataV1>;
  ICedarSFTMetadataV1?: FeatureInterface<ICedarSFTMetadataV1>;
  ICedarMetadataV1?: FeatureInterface<ICedarMetadataV1>;
  IPublicMetadataV0?: FeatureInterface<IPublicMetadataV0>;
  IRoyaltyV0?: FeatureInterface<IRoyaltyV0>;
  IPublicRoyaltyV0?: FeatureInterface<IPublicRoyaltyV0>;
  IPrimarySaleV0?: FeatureInterface<IPrimarySaleV0>;
  IPrimarySaleV1?: FeatureInterface<IPrimarySaleV1>;
  IPublicPrimarySaleV1?: FeatureInterface<IPublicPrimarySaleV1>;
  ICedarSplitPaymentV0?: FeatureInterface<ICedarSplitPaymentV0>;
  IERC721V0?: FeatureInterface<IERC721V0>;
  INFTSupplyV0?: FeatureInterface<INFTSupplyV0>;
  IERC1155V0?: FeatureInterface<IERC1155V0>;
  IERC1155SupplyV0?: FeatureInterface<IERC1155SupplyV0>;
  IERC1155SupplyV1?: FeatureInterface<IERC1155SupplyV1>;
  ISFTSupplyV0?: FeatureInterface<ISFTSupplyV0>;
  IERC2981V0?: FeatureInterface<IERC2981V0>;
  IPublicOwnableV0?: FeatureInterface<IPublicOwnableV0>;
};

export type SupportedInterfaces =
  | ICedarFeaturesV0
  | ICedarVersionedV0
  | ICedarVersionedV1
  | ICedarVersionedV2
  | IMulticallableV0
  | ICedarAgreementV0
  | ICedarAgreementV1
  | IPublicAgreementV0
  | IPublicAgreementV1
  | ICedarUpdateBaseURIV0
  | ICedarUpgradeBaseURIV0
  | IPublicUpdateBaseURIV0
  | ICedarNFTIssuanceV0
  | ICedarNFTIssuanceV1
  | ICedarNFTIssuanceV2
  | ICedarNFTIssuanceV3
  | ICedarNFTIssuanceV4
  | IPublicNFTIssuanceV0
  | IPublicNFTIssuanceV1
  | ICedarSFTIssuanceV0
  | ICedarSFTIssuanceV1
  | ICedarSFTIssuanceV2
  | ICedarSFTIssuanceV3
  | IPublicSFTIssuanceV0
  | IPublicSFTIssuanceV1
  | ICedarPausableV0
  | ICedarPremintV0
  | ICedarLazyMintV0
  | ICedarNFTMetadataV1
  | ICedarSFTMetadataV1
  | ICedarMetadataV1
  | IPublicMetadataV0
  | IRoyaltyV0
  | IPublicRoyaltyV0
  | IPrimarySaleV0
  | IPrimarySaleV1
  | IPublicPrimarySaleV1
  | ICedarSplitPaymentV0
  | IERC721V0
  | INFTSupplyV0
  | IERC1155V0
  | IERC1155SupplyV0
  | IERC1155SupplyV1
  | ISFTSupplyV0
  | IERC2981V0
  | IPublicOwnableV0;

export interface FeatureInterfaceFactory<T> {
  connect(address: string, signerOrProvider: Signer | Provider): T;
}

export type FeatureInterfaceFactoryMapper = {
  [key: string]: FeatureInterfaceFactory<SupportedInterfaces>;
};

export type TokenStandard = 'ERC721' | 'ERC1155' | 'ERC20';
// ...
export type CollectionInfo = {
  chainId: ChainId;
  address: Address;
  tokenStandard: TokenStandard | null;
};

export type CollectionCallData = {
  method: string;
  args?: { [key: string]: unknown };
  signer: Signerish;
  supportedFeatures: string[];
};

export type DebugHandler = (collection: CollectionInfo, message: string, ...optionalParams: unknown[]) => void;

export type ErrorHandler = (
  message: string,
  error: Error,
  collection: CollectionInfo,
  callData: CollectionCallData,
) => void;